Update `combine_bytes` usage
The `combine_bytes` macro is really just there to reconstruct a byte
array into a single number. The CRC byte for commands is stored in the
last byte of the command buffer so `0..cmd_buffer.len() - 1`, should be
`0..cmd_buffer.len() - 2` to not include the CRC byte in the combined
result.