Fix address checks for clockless read/write
Originally wrote `spi_read_register` to check if address was less than
0x30. That was incorrect it should check if less than 0xff. The
`spi_writ_register` method should be checking if address is less than
0x30.

In the Atmel driver the read/write register functions do a clockless
read/write if the address is under a certain value. It's a clockless
read if `address < 0xff` and a clockless write if `address < 0x30`. I
have not seen any addresses that are smaller than 0xff or 0x30, but I
put them in just in case.