Replace spi transfer with spi send/read
The transfer method takes a slice of data, sends a byte then receives
a byte until all bytes in the slice have been sent. This is fine,
but the Atwinc1500 doesn't send a response until a full command has
been received. If the command is
[0xC9, 0x0, 0xE8, 0x24, 0x0, 0x0, 0x0, 0x52, 0x5C], then the result
will be [0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0]. The transfer
method replaces bytes sent with bytes read and in this case there are
no more bytes to read/store when the command is entirely sent. The
command buffer either needs to be larger or the read/send process
needs to write the read bytes to the beginning of the slice instead.