Remove cmd_buffer parameter from most spi methods
The read/write register and read/write data methods in SpiLayer all
had cmd_buffer parameters. The method calling on of the SpiLayer
methods should not be concerned with creating the array to hold their
data. Instead, the SpiLayer methods should define the arrays, do work
on the arrays and then extract and return the result (data(u32)/Ok()).

All of the spots where these methods are called were fixed too. The
disable_crc method now sends the write directly with spi_command
because there is no way to insert the CRC byte otherwise.