Fix connect_network method
Swapped two bytes around and now the atwinc is connecting to networks.
One thing I'll have to be mindful of is the newer firmwares added a
new connect command so I may want to support both. This verifies that
the spi write_data method and the hif send method are working.